feat(x/ecocredit): default values for new ORM params
